refactor: 영화 관련 패키지명 변경
