+++ app/src/main/java/woowacourse/movie/presentation/movie/MovieListContract.kt
@@ -1,7 +1,7 @@
-package woowacourse.movie.presentation.movies
+package woowacourse.movie.presentation.movie
 
 import woowacourse.movie.domain.model.Movie
-import woowacourse.movie.presentation.movies.adapter.MovieListItem
+import woowacourse.movie.presentation.movie.adapter.MovieListItem
 
 interface MovieListContract {
     interface View {
