+++ app/src/main/java/woowacourse/movie/presentation/movie/MovieListFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.movies
+package woowacourse.movie.presentation.movie
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,9 +7,9 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import woowacourse.movie.databinding.FragmentMovieListBinding
 import woowacourse.movie.domain.model.Movie
-import woowacourse.movie.presentation.movies.adapter.MovieAdapter
-import woowacourse.movie.presentation.movies.adapter.MovieListClickListener
-import woowacourse.movie.presentation.movies.adapter.MovieListItem
+import woowacourse.movie.presentation.movie.adapter.MovieAdapter
+import woowacourse.movie.presentation.movie.adapter.MovieListClickListener
+import woowacourse.movie.presentation.movie.adapter.MovieListItem
 import woowacourse.movie.presentation.theater.TheaterFragment
 
 class MovieListFragment :
