+++ app/src/main/java/woowacourse/movie/presentation/movie/adapter/MovieAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.movies.adapter
+package woowacourse.movie.presentation.movie.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
