+++ app/src/main/java/woowacourse/movie/presentation/movie/adapter/MovieListClickListener.kt
@@ -1,3 +1,3 @@
-package woowacourse.movie.presentation.movies.adapter
+package woowacourse.movie.presentation.movie.adapter
 
 interface MovieListClickListener : MovieViewHolder.ClickListener
