+++ app/src/main/java/woowacourse/movie/presentation/movie/adapter/MovieListItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.movies.adapter
+package woowacourse.movie.presentation.movie.adapter
 
 import androidx.annotation.DrawableRes
 import woowacourse.movie.R
